Add tests

+++ sqrl-math/src/main/java/com/datasqrl/math/distribution/BinomialDistribution.java
@@ -6,9 +6,9 @@ import org.apache.flink.table.functions.ScalarFunction;
 /** Calculates the probability mass function for a binomial distribution. */
 @AutoService(ScalarFunction.class)
 public class BinomialDistribution extends ScalarFunction {
-    public Double eval(Integer trials, Double probability, Integer x) {
+    public Double eval(Long trials, Double probability, Long x) {
         if (trials == null || probability == null || x == null) return null;
-        org.apache.commons.math3.distribution.BinomialDistribution dist = new org.apache.commons.math3.distribution.BinomialDistribution(trials, probability);
-        return dist.probability(x);
+        org.apache.commons.math3.distribution.BinomialDistribution dist = new org.apache.commons.math3.distribution.BinomialDistribution(trials.intValue(), probability);
+        return dist.probability(x.intValue());
     }
 }
