+++ sqrl-math/src/main/java/com/datasqrl/math/distribution/PoissonDistribution.java
@@ -6,9 +6,9 @@ import org.apache.flink.table.functions.ScalarFunction;
 /** Calculates the cumulative probability for a Poisson distribution. */
 @AutoService(ScalarFunction.class)
 public class PoissonDistribution extends ScalarFunction {
-    public Double eval(Double mean, Integer x) {
+    public Double eval(Double mean, Long x) {
         if (mean == null || x == null) return null;
         org.apache.commons.math3.distribution.PoissonDistribution dist = new org.apache.commons.math3.distribution.PoissonDistribution(mean);
-        return dist.cumulativeProbability(x);
+        return dist.cumulativeProbability(x.intValue());
     }
 }
