+++ sqrl-math/src/main/java/com/datasqrl/math/util/Scalb.java
@@ -7,8 +7,8 @@ import org.apache.flink.table.functions.ScalarFunction;
 /** Multiplies x by 2^scaleFactor. */
 @AutoService(ScalarFunction.class)
 public class Scalb extends ScalarFunction {
-    public Double eval(Double x, Integer scaleFactor) {
+    public Double eval(Double x, Long scaleFactor) {
         if (x == null || scaleFactor == null) return null;
-        return FastMath.scalb(x, scaleFactor);
+        return FastMath.scalb(x, scaleFactor.intValue());
     }
 }
